+++ vaadin-maps-leaflet-flow/src/main/java/software/xdev/vaadin/maps/leaflet/layer/vector/LRectangle.java
@@ -22,6 +22,7 @@ import software.xdev.vaadin.maps.leaflet.registry.LComponentManagementRegistry;
 /**
  * @see <a href="https://leafletjs.com/reference.html#rectangle">Leaflet docs</a>
  */
+@SuppressWarnings("java:S110")
 public class LRectangle extends LPolyline
 {
 	public LRectangle(